requirements.txt
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyGithub==1.54
PyJWT==1.7.1
python-dotenv==0.15.0
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
